requirements.txt
jupyter>=1.0.0
pytorch>=1.8.1
torchvision>=0.9.1
matplotlib>=3.3.4